% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dcca}
\alias{dcca}
\title{Detrended Cross-Correlation Analysis}
\usage{
dcca(x, y, order, scales)
}
\arguments{
\item{x}{Vetor numérico (série temporal).}

\item{y}{Vetor numérico (série temporal).}

\item{order}{Ordem do polinômio para detrending (ex: 1=linear).}

\item{scales}{Vetor de inteiros com as escalas para o cálculo.}
}
\value{
Uma lista com \code{scales} e os coeficientes \code{rho}.
}
\description{
Detrended Cross-Correlation Analysis
}
