% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{detrend_cov}
\alias{detrend_cov}
\title{Detrended Covariance
Wrapper function for the rust function for doing the detrending.}
\usage{
detrend_cov(x, y, m)
}
\arguments{
\item{x}{numeric vector.}

\item{y}{numeric vector.}

\item{m}{polinimial order for detrending (ex: 1=linear).}
}
\value{
vector with the sum of squared residuals.
}
\description{
Detrended Covariance
Wrapper function for the rust function for doing the detrending.
}
