% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{logscale}
\alias{logscale}
\title{Create logarithmically spaced scales.}
\usage{
logscale(scale_min, scale_max, scale_ratio)
}
\description{
This generates a vector of scales starting from \code{scale_min}, where each
subsequent scale is the previous one multiplied by \code{scale_ratio}.
The process stops once \code{scale_max} is exceeded.
}
\section{Arguments}{
\itemize{
\item \code{scale_min} - An integer indicating the minimum scale to be resolved.
\item \code{scale_max} - An integer indicating the maximum scale to be resolved.
\item \code{scale_ratio} - A float indicating the ratio for successive scales.
}
}

\section{Returns}{
A \verb{Vec<i32>} of logarithmically spaced scales, sorted and unique.
}

\section{Examples}{
// This would be called from R like so:
// scales <- logscale(scale_min = 16, scale_max = 1024, scale_ratio = 2)
}

