% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhodcca.R
\name{rhodcca}
\alias{rhodcca}
\title{rhodcca}
\usage{
rhodcca(
  series1,
  series2,
  scale_min = NULL,
  scale_max = NULL,
  order = 1,
  scale_ratio = 1.05
)
}
\arguments{
\item{series1}{numeric vector}

\item{series2}{numeric vector}

\item{scale_min}{an integer indicating the minimum scale to be resovled}

\item{scale_max}{an integer indicating the maximum scale to be resovled}

\item{order}{is an integer indicating the polynomial order used for
detrending the local windows (e.g, 1 = linear, 2 = quadratic, etc.).}

\item{scale_ratio}{a double indicating the ratio by which scale successive
scales. For example, scale_ratio = 2 would create a scales increasing by
a power of 2.}
}
\value{
A list of dcca.out (dataframe with columns scales, rho) and error
}
\description{
Calculate rhodcca
}
\examples{
rhodcca.out <- rhodcca(series1, series2)

}
